add calculate rating after create rating

+++ db/migrate/20170924164215_add_rate_to_user.rb
@@ -1,5 +1,5 @@
 class AddRateToUser < ActiveRecord::Migration[5.1]
   def change
-    add_column :users, :rate, :float, precision: 5, scale: 2, default: 0
+    add_column :users, :rate, :decimal, precision: 5, scale: 2, default: 0
   end
 end
